refactors common.ts by moving it and external.ts to cli/common, renaming as index.ts, refactoring tests, and updating references | Daniel
