+++ packages/cli/src/__tests__/GuidedInstall/runner.test.ts
@@ -2,17 +2,17 @@
  * © 2021 Thoughtworks, Inc.
  */
 
-import * as common from '../../GuidedInstall/common'
-import * as external from '../../GuidedInstall/external'
+import * as common from '../../common'
+import * as external from '../../common/external'
 import { GuidedInstall } from '../../GuidedInstall/runner'
 
-jest.mock('../../GuidedInstall/common')
+jest.mock('../../common')
 const mockListPrompt: jest.Mock = common.listPrompt as jest.Mock
 const mockConfirmPrompt: jest.Mock = common.confirmPrompt as jest.Mock
 const mockInputPrompt: jest.Mock = common.inputPrompt as jest.Mock
 const mockCreateEnvFile: jest.Mock = common.createEnvFile as jest.Mock
 
-jest.mock('../../GuidedInstall/external')
+jest.mock('../../common/external')
 const mockLog: jest.Mock = external.log as jest.Mock
 const mockRunCmd: jest.Mock = external.runCmd as jest.Mock
