+++ packages/cli/src/common/external.ts
